SliderTraining-Klein/requirements.txt
torch>=2.1.0
einops>=0.7.0
safetensors>=0.4.0
transformers>=4.45.0
huggingface-hub>=0.34.0
torchvision>=0.16.0
Pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
omegaconf>=2.3.0
opencv-python>=4.9.0
datasets>=2.18.0
